requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-crispy-forms==1.14.0
django-dynamic-preferences==1.14.0
django-render-block==0.9.1
django-split-settings==1.2.0
django-utils-six==2.0
html2text==2020.1.16
idna==3.4
persisting-theory==1.0
Pillow==9.2.0
python-decouple==3.6
pytz==2022.7.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
